リファクタその１
