+++ backend/app/controllers/api/v1/admin/discussion_threads_admin_recommended_topic_controller.rb
@@ -6,25 +6,16 @@ module Api
               resentThreads = DiscussionThreadAdminRecommended.fetch_recommended
               render_json_response(resentThreads)
           end
+
           def bulk_add
             DiscussionThreadAdminRecommended.bulk_add(params[:selectedThreads])
-            Rails.logger.debug("bulk_add: #{params[:selectedThreads]}")
             render_json_response({ message: "おすすめスレッドが追加されました。" })
           end
+
           def bulk_delete
             DiscussionThreadAdminRecommended.bulk_delete(params[:selectedThreads])
-            Rails.logger.debug("bulk_delete: #{params[:selectedThreads]}")
             render_json_response({ message: "おすすめスレッドが削除されました。" })
           end
-          # def destroy
-          #   resentThreads = DiscussionThreadAdmin.delete_admin(params[:id])
-          #   render_json_response(resentThreads)
-          # end
-          # def update
-          #   Rails.logger.debug("update_title_admin: #{params[:id]}")
-          #   resentThreads = DiscussionThreadAdmin.update_title_admin(params[:id],params[:thread_title])
-          #   render_json_response(resentThreads)
-          # end
       end
     end
   end
