--- backend/app/controllers/api/v1/admin/posts_report_controller.rb
@@ -1,12 +1,0 @@
-module Api
-  module V1
-    module Admin
-      class PostsReportController < ApplicationController
-        def index
-          posts = Post.reported
-          render json: { data: posts }, status: :ok
-        end
-      end
-    end
-  end
-end
