+++ frontend/.env
@@ -4,13 +4,16 @@ VITE_DISCUSSION_THREAD_POPULAR_API_URL="http://localhost:3000/api/v1/discussion_
 VITE_DISCUSSION_THREAD_POPULAR_API_URL="http://localhost:3000/api/v1/discussion_threads_week_populartopic"
 VITE_DISCUSSION_THREAD_RECENT_API_URL="http://localhost:3000/api/v1/discussion_threads_newtopic"
 VITE_DISCUSSION_THREAD_RECOMMEND_API_URL="http://localhost:3000/api/v1/discussion_threads_recommendtopic"
+VITE_POSTS_API_URL="http://localhost:3000/api/v1/posts"
+
+//report api
+VITE_POST_REPORT_API_URL="http://localhost:3000/api/v1/report/posts_report"
+VITE_ADMIN_POST_REPORT_API_URL="http://localhost:3000/api/v1/report/posts_report"
+VITE_REPORT_CONTENTS_API_URL="http://localhost:3000/api/v1/report/report_contents"
+
+//admin api
+VITE_LOGIN_ADMIN_USER_API_URL="http://localhost:3000/api/v1/admin/admin"
+VITE_LOGIN_API_URL="http://localhost:3000/api/v1/admin/login"
 VITE_DISCUSSION_THREAD_ADMIN_API_URL="http://localhost:3000/api/v1/admin/discussion_threads_admin_topic"
 VITE_DISCUSSION_THREAD_ADMIN_RECOMMENDED_API_URL="http://localhost:3000/api/v1/admin/discussion_threads_admin_recommended_topic"
 VITE_DISCUSSION_THREAD_ADMIN_POSTS_API_URL="http://localhost:3000/api/v1/admin/discussion_threads_admin_posts"
-VITE_POSTS_API_URL="http://localhost:3000/api/v1/posts"
-VITE_POST_REPORT_API_URL="http://localhost:3000/api/v1/report/posts_report"
-VITE_POSTS_API_URL="http://localhost:3000/api/v1/posts"
-VITE_LOGIN_ADMIN_USER_API_URL="http://localhost:3000/api/v1/admin"
-VITE_LOGIN_API_URL="http://localhost:3000/api/v1/login"
-VITE_ADMIN_POST_REPORT_API_URL="http://localhost:3000/api/v1/admin/posts_report"
-VITE_REPORT_CONTENTS_API_URL="http://localhost:3000/api/v1/admin/report_contents"
