+++ frontend/src/components/Admin/Dashboard/sideBar/sideBar.module.css
@@ -74,3 +74,28 @@
 .columContainer p {
     color: white;
 }
+
+.logoutWrapper {
+    margin-top: auto;
+    padding: 20px;
+    display: flex;
+    justify-content: center;
+}
+
+.logoutButton {
+    background: #ff4d4d;
+    color: white;
+    font-weight: bold;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 30px;
+    font-size: 16px;
+    cursor: pointer;
+    transition: 0.3s ease;
+    box-shadow: 0 4px 10px rgba(255, 77, 77, 0.4);
+}
+
+.logoutButton:hover {
+    background: #ff1a1a;
+    transform: scale(1.05);
+}
